requirements.txt
streamlit
requests
duckduckgo-search
wikipedia-api
yfinance
python-dotenv
google-generativeai
python-multipart